+++ xla/backends/cpu/testlib/kernel_runner_extension.cc
@@ -60,13 +60,13 @@ void ImportBaseClasses(const nb::module_& kernel_runner_module) {
   nb::module_::import_(absl::StrCat(xla_module, ".codegen.testlib").c_str());
 }
 
-NB_MODULE(_extention, kernel_runner_module) {
+NB_MODULE(_extension, kernel_runner_module) {
   // We depend on the base classes so must import them before python tries to
   // register the derived versions.
   ImportBaseClasses(kernel_runner_module);
 
-  nb::class_<LlvmIrKernelSpec, KernelSpec>(kernel_runner_module,
-                                           "LlvmIrKernelSpec");
+  nb::class_<LlvmIrKernelSpec, KernelSpec> give_me_a_name(kernel_runner_module,
+                                                          "LlvmIrKernelSpec");
 
   // Use a tuple and cast to ThreadDim to take advantage of built in bindings.
   using NbThreadDim = std::tuple<uint64_t, uint64_t, uint64_t>;